Arreglado Bug de dupla
Habia un bug que hacia que en la interfaz se mostraran 2 espacios del mismo casillero
